Checking Betty

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 void print_element(const list_t *h)
 {
 	if (h->str != '\0')
-		printf("[%u] %s\n", h->len, h->str);
+		printf("[%d] %s\n", h->len, h->str);
 	else
 		printf("[0] (nil)\n");
 }
@@ -20,7 +20,7 @@ void print_element(const list_t *h)
  */
 size_t print_list(const list_t *h)
 {
-	size_t i = 0;
+	int i = 0;
 
 	while (h->next != NULL)
 	{
